split numpy test
