+++ .github/workflows/python.yaml
@@ -72,7 +72,7 @@ jobs:
       matrix:
         os: ["ubuntu-latest", "macos-13", "windows-latest"]
         python-version: ["3.9", "3.10", "3.11"]
-        module: ["xorbits", "kubernetes"]
+        module: ["xorbits", "xorbits/numpy", "xorbits/pandas", "kubernetes"]
         exclude:
           - { os: macos-13, python-version: 3.10}
           - { os: macos-13, python-version: 3.9}
@@ -82,8 +82,6 @@ jobs:
           - { os: macos-13, module: kubernetes}
         include:
           - { os: ubuntu-latest, module: _mars/dataframe, python-version: 3.9 }
-          - { os: ubuntu-latest, module: _mars/tensor, python-version: 3.9 }
-          - { os: ubuntu-latest, module: _mars/learn, python-version: 3.9 }
           - { os: ubuntu-latest, module: learn, python-version: 3.9 }
           - { os: ubuntu-latest, module: mars-core, python-version: 3.9 }
           - { os: ubuntu-20.04, module: hadoop, python-version: 3.9 }
@@ -135,10 +133,10 @@ jobs:
         PYTHON: ${{ matrix.python-version }}
       if: ${{ matrix.module != 'gpu' }}
       run: |
-        pip install -e "git+https://github.com/frostyplanet/xoscar.git@fix_0.0.0.0#subdirectory=python&egg=xoscar"
+        pip install -e "git+https://github.com/xorbitsai/xoscar.git@main#subdirectory=python&egg=xoscar"
         pip install "numpy<2.0.0" scipy cython pyftpdlib coverage flaky numexpr
 
-        if [[ "$MODULE" == "xorbits" ]]; then
+        if [[ "$MODULE" == "xorbits" || "$MODULE" == "numpy" ]]; then
           pip install openpyxl
         fi
         if [[ "$MODULE" == "mars-core" ]]; then
@@ -281,10 +279,16 @@ jobs:
         MODULE: ${{ matrix.module }}
       run: |
         if [[ "$MODULE" == "xorbits" ]]; then
-          pytest --ignore xorbits/_mars/ --ignore xorbits/xgboost --ignore xorbits/lightgbm \
+          pytest --ignore xorbits/_mars/ --ignore xorbits/pandas --ignore xorbits/numpy \
+            --ignore xorbits/xgboost --ignore xorbits/lightgbm \
             --ignore xorbits/datasets --timeout=1500 \
             -W ignore::PendingDeprecationWarning \
             --cov-config=setup.cfg --cov-report=xml --cov=xorbits xorbits
+        elif [[ "$MODULE" == "xorbits.pandas" ]]; then
+          pytest --timeout=1500 \
+            -W ignore::PendingDeprecationWarning \
+            --cov-config=setup.cfg --cov-report=xml \
+            --cov=xorbits xorbits/pandas
         elif [[ "$MODULE" == "mars-core" ]]; then
           pytest --forked --log-level=DEBUG --ignore xorbits/_mars/dataframe --ignore xorbits/_mars/tensor \
             --ignore xorbits/_mars/learn --ignore xorbits/_mars/remote \
